File changes

+++ src/components/application/pdf/SPS902Document.jsx
@@ -178,7 +178,7 @@ export default function SPS902Document({ applicant, formData = {} }) {
                                 <div>
                                     <span>ได้รับค่าจ้างงวดสุดท้ายเมื่อเดือน</span>
                                     <span className={`border-b border-dotted border-slate-400 inline-block min-w-[40px] text-center px-2 pb-1 mx-1`} style={{ verticalAlign: 'baseline', minHeight: '1.2em' }}>
-                                        {formData.lastSalaryMonth ? new Date(formData.lastSalaryMonth).toLocaleDateString('th-TH', { month: 'long' }) : '\u00A0'}
+                                        {formData.lastSalaryMonth ? new Date(formData.lastSalaryMonth).toLocaleDateString('th-TH', { month: 'short' }) : '\u00A0'}
                                     </span>
                                     <span>พ.ศ.</span>
                                     <span className={`border-b border-dotted border-slate-400 inline-block min-w-[40px] text-center px-2 pb-1 mx-1`} style={{ verticalAlign: 'baseline', minHeight: '1.2em' }}>
